backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-telegram-bot==21.3
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
google-generativeai==0.7.2
python-dotenv==1.0.1
httpx==0.27.0
asyncpg==0.29.0
pydantic==2.7.1
aiohttp==3.9.5